# openapi_schema_for_custom_gpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RIGA Angel One Backend",
    "version": "1.0.0",
    "description": "Read-only Angel One market data API for RIGA Custom GPT."
  },
  "servers": [
    {
      "url": "https://YOUR-RENDER-URL.onrender.com"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "checkHealth",
        "summary": "Check backend health",
        "responses": {
          "200": {
            "description": "Backend is running"
          }
        }
      }
    },
    "/ltp": {
      "get": {
        "operationId": "getLtp",
        "summary": "Get LTP from Angel One",
        "parameters": [
          {
            "name": "exchange",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tradingsymbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "symboltoken",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "LTP response"
          }
        }
      }
    },
    "/riga-signal-demo": {
      "get": {
        "operationId": "getRigaSignalDemo",
        "summary": "Demo RIGA signal endpoint",
        "parameters": [
          {
            "name": "exchange",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tradingsymbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "symboltoken",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Demo signal response"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}